move plan table into home
