+++ src/interface/src/app/home/home.component.html
@@ -1,1 +1,2 @@
 <app-region-selection></app-region-selection>
+<app-plan-table></app-plan-table>
