+++ src/interface/src/app/home/plan-table/delete-plan-dialog/delete-plan-dialog.component.scss
