+++ src/interface/src/app/home/plan-table/plan-table.component.scss
